+++ changelog.md
@@ -1,5 +1,6 @@
 latest
  - Amplitude threshold condense selected action, enter an upper and lower amplitude value (between 0 and 1), the sample will be modified to only include amplitudes within the given range, gaps between the audio is removed condensing the sample size.
+ - Ctrl + Click on Filename column header to sort the list by numeric filename content only.
 
 2024.05.14 1.4.14
  - Added basic error catch in the joinAll method to help surface chain build errors.