+++ changelog.md
@@ -1,4 +1,4 @@
-latest
+2024.05.28 1.4.15
  - Amplitude threshold condense selected action, enter an upper and lower amplitude value (between 0 and 1), the sample will be modified to only include amplitudes within the given range, gaps between the audio is removed condensing the sample size.
  - Ctrl + Click on Filename column header to sort the list by numeric filename content only.
  - Improved drop file order consistency.